requirements.txt
libscrc==1.7.1
Pillow==7.2.0
qrcode==6.1